pyproject.toml
[tool.poetry]
name = "chat2dutch"
version = "0.1.0"
description = "This is an app to help people learn Dutch"
authors = ["Xin Wang <xinaiarrival@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gradio = "^4.44.0"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
openai = "^1.44.1"
sqlalchemy = "^2.0.34"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
wiktionaryparser = "^0.0.97"
wikipedia-api = "^0.7.1"
unidecode = "^1.3.8"
deep-translator = "^1.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
